feat(cxx): Split the runtime header to many small headers
