+++ Jikka.cabal
@@ -21,9 +21,16 @@ extra-source-files:
     README.md
     CHANGELOG.md
 data-files:
-    runtime/include/jikka/base.hpp
     runtime/include/jikka/convex_hull_trick.hpp
+    runtime/include/jikka/divmod.hpp
+    runtime/include/jikka/error.hpp
+    runtime/include/jikka/matrix.hpp
+    runtime/include/jikka/modulo.hpp
+    runtime/include/jikka/modulo_matrix.hpp
+    runtime/include/jikka/not_modulo.hpp
+    runtime/include/jikka/range.hpp
     runtime/include/jikka/segment_tree.hpp
+    runtime/include/jikka/slope_trick.hpp
 
 source-repository head
   type: git
